program that finds and prints the first 98 Fibonacci numbers, starting with 1 and 2, followed by a new line

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,13 +7,13 @@
 int main(void)
 {
 	int n = 0;
-	long j = 1, k = 2;
+	unsigned long long j = 1, k = 2;
 
 	printf("%ld, %ld ", j, k);
 	n = 2;
 	while (n < 98)
 	{
-		long i = j + k;
+		unsigned long long i = j + k;
 
 		printf("%ld", i);
 		n++;
